backend-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
groq==0.11.0

# Alternative AI providers (uncomment if needed):
# google-generativeai==0.3.2  # For Google Gemini
# openai==1.10.0              # For OpenAI GPT
# anthropic==0.18.0           # For Claude
